requirements.txt
altair>=5.0.0
pandas
python-dotenv
streamlit
plotly
scikit-learn
yfinance
google-generativeai
google-cloud-bigquery
# Placeholder for Gemini 1.5 integration
openai  # or replace with official Gemini client when available
pandas-gbq
tabulate
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-requests
